src/requirements.txt
black==22.10.0
Django==4.2.6
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-phonenumber-field==7.3.0
django-taggit==5.0.1
djoser==2.2.2
drf-yasg==1.21.7
factory_boy==3.3.0
Markdown==3.5
Pillow==10.1.0
phonenumbers==8.13.29
pytest==7.4.3
pytest-django==4.7.0
python-decouple==3.8
requests==2.31.0
social-auth-app-django==5.4.0
Unidecode==1.3.7
